feat: split Api.js, do not copy node_modules, & fix tests
