+++ lib/create.js
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 /*
     Licensed to the Apache Software Foundation (ASF) under one
     or more contributor license agreements.  See the NOTICE file
@@ -26,25 +24,25 @@ const events = require('cordova-common').events;
 const check_reqs = require('./check_reqs');
 
 // exported method to create a project, returns a promise that resolves with null
-module.exports.createProject = function (project_path, package_name, project_name) {
-/*
-    // create the dest and the standard place for our api to live
-    // platforms/platformName/cordova/Api.js
-*/
+module.exports.createProject = function (project_path, package_name, project_name, opts, root_config) {
+    opts = opts || {};
 
-    events.emit('log', 'Creating Cordova project for cordova-browser:');
-    events.emit('log', '\tPath: ' + project_path);
-    events.emit('log', '\tName: ' + project_name);
-
-    // Set default values for path, package and name
-    project_path = project_path || 'CordovaExample';
+    project_path = path.relative(process.cwd(), project_path);
 
     // Check if project already exists
     if (fs.existsSync(project_path)) {
-        events.emit('error', 'Oops, destination already exists! Delete it and try again');
-    } else {
-        fs.mkdirSync(project_path);
+        throw new CordovaError('Project already exists');
     }
+
+    package_name = package_name || 'org.apache.cordova.hellocordova';
+    project_name = project_name || 'Hello Cordova';
+
+    events.emit('log', 'Creating Cordova project for the browser platform:');
+    events.emit('log', `\tPath: ${project_path}`);
+    events.emit('log', `\tPackage: ${package_name}`);
+    events.emit('log', `\tName: ${project_name}`);
+
+    fs.mkdirSync(project_path);
 
     // Check that requirements are met and proper targets are installed
     if (!check_reqs.run()) {
@@ -52,21 +50,8 @@ module.exports.createProject = function (project_path, package_name, project_nam
         events.emit('error', 'Please make sure you meet the software requirements in order to build a browser cordova project');
     }
 
-    // copy template/* to project_path directory ( recursive )
+    // copy templates/* to project_path directory ( recursive )
     fs.cpSync(path.join(ROOT, 'templates'), project_path, { recursive: true });
-
-    // TODO: stop copying cordova-browser logic into platforms/browser/
-    // recreate our node_modules structure in the new project
-    const nodeModulesDir = path.join(ROOT, 'node_modules');
-    if (fs.existsSync(nodeModulesDir)) {
-        fs.cpSync(nodeModulesDir, path.join(project_path, 'cordova/node_modules'), { recursive: true });
-    }
-
-    // copy check_reqs file
-    fs.cpSync(
-        path.join(ROOT, 'lib/check_reqs.js'),
-        path.join(project_path, 'cordova/lib/check_reqs.js')
-    );
 
     // create platform_www dir if it does not exist yet
     const platform_www = path.join(project_path, 'platform_www');
